Project Update. Implement orderTotal() Method for TF_Order. Implement getter getPrice() for TF_Product. Implement getter getSize() for TF_Vector. Overload [] operator for TF_Vector.

+++ Headers/TF_Order.hpp
@@ -17,7 +17,7 @@ public:
 
     friend std::ostream &operator<<(std::ostream &out, const TF_Order &object);
 
-    double orderTotal();
+    int orderTotal();
     double addCharge(double charge);
 };
